202204012108 - Commit
Fixed a lot of little references to raw pointers in main, Monster, Player, and Object files.

error202204012108.rtf

(Player.cpp) In the output stream operator overload, I had to modify the loop to better handle the unique pointer. When switching over to a unique pointer from raw pointers it was throwing a nasty error. So, I refactored it.

Game compiles, runs, and shows no errors...